Coffee Machine
This Python program simulates a coffee machine. It takes user input to select a coffee, checks if there are enough resources to make the coffee, processes the payment, and then makes the coffee if all conditions are met.

Functions:
is_resource_sufficient(order_ingredients):

Checks if there are enough ingredients in the machine to make the selected coffee. If not, it prints a message indicating which ingredient is insufficient and returns False.

process_coins():

Prompts the user to insert coins (quarters, dimes, nickels, pennies) and calculates the total amount inserted.

is_transaction_successful(money_received, drink_cost):

Checks if the inserted money is enough to cover the cost of the selected drink. If the payment is sufficient, it returns change to the user, updates the machine’s profit, and returns True. Otherwise, it refunds the money and returns False.

make_coffee(drink_name, order_ingredients):

Deducts the required ingredients from the available resources and prints a message indicating that the coffee is ready.